Minor refactor and docstring
