+++ iotile_ext_cloud/iotile/cloud/cloud.py
@@ -5,7 +5,6 @@ from io import BytesIO
 import requests
 import getpass
 import time
-from calendar import timegm
 
 from iotile_cloud.api.connection import Api
 from iotile_cloud.api.exceptions import RestHttpBaseException, HttpNotFoundError
@@ -13,17 +12,7 @@ from iotile.core.dev.registry import ComponentRegistry
 from iotile.core.dev.config import ConfigManager
 from iotile.core.exceptions import ArgumentError, ExternalError
 from iotile.core.utilities.typedargs import context, param, return_type, annotated, type_system
-
-def get_timestamp(utcdate):
-    """ retrieves the UNIX UTC timestamp from formatted text ( 06 Sep 2017 16:50:46 )"""
-    months = {'Jan': 1, 'Feb': 2, 'Mar': 3, 'Apr': 4, 'May': 5, 'Jun': 6,
-              'Jul': 7, 'Aug': 8, 'Sep': 9, 'Oct': 10, 'Nov': 11, 'Dec': 12}
-    utcdate = utcdate.split(' ')
-    year, day = utcdate[2], utcdate[0]
-    month = months[utcdate[1]]
-    utcdate = utcdate[-1].split(':')
-    hour, m, sec = utcdate[0], utcdate[1], utcdate[2]
-    return(timegm(list(map(lambda x: int(x),(year,month,day,hour,m,sec)))))
+from iotile.cloud.utilities import get_timestamp
 
 @context("IOTileCloud")
 class IOTileCloud(object):
@@ -102,6 +91,8 @@ class IOTileCloud(object):
 
     @return_type("bool")
     def check_time(self):
+        """Checks if the current system time is consistent with iotile.cloud time
+        """
         cloud_time = get_timestamp(requests.get('https://iotile.cloud').headers['Date'][5:-4])
         curtime = time.time()
         return (curtime < cloud_time + 300 and curtime > cloud_time - 300)
